core/sphinx: remove old shitty benchmark tests
